PR comments incorporated
