+++ test/Microsoft.AspNetCore.Mvc.Core.Test/AcceptedResultTests.cs
@@ -9,18 +9,16 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc.Abstractions;
 using Microsoft.AspNetCore.Mvc.Formatters;
 using Microsoft.AspNetCore.Mvc.Internal;
-using Microsoft.AspNetCore.Mvc.Routing;
 using Microsoft.AspNetCore.Routing;
-using Microsoft.AspNetCore.Testing;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging.Testing;
 using Moq;
 using Newtonsoft.Json;
 using Xunit;
 
-namespace Microsoft.AspNetCore.Mvc
+namespace Microsoft.AspNetCore.Mvc.Core.Test
 {
-    public class AcceptedAtActionResultTests
+    public class AcceptedResultTests
     {
         public static TheoryData<object> ValuesData
         {
@@ -41,16 +39,10 @@ namespace Microsoft.AspNetCore.Mvc
 
         [Theory]
         [MemberData(nameof(ValuesData))]
-        public void AcceptedAtActionResult_InitializesStatusCodeAndValue(object value)
+        public void AcceptedObjectResult_InitializesStatusCodeAndValue(object value)
         {
-            // Arrange 
-            var expectedUrl = "testAction";
-
-            //Act         
-            var result = new AcceptedAtActionResult(
-                actionName: expectedUrl,
-                controllerName: null,
-                routeValues: null, value: value);
+            // Arrange & Act
+            var result = new AcceptedResult("testlocation", value);
 
             // Assert
             Assert.Equal(StatusCodes.Status202Accepted, result.StatusCode);
@@ -59,22 +51,15 @@ namespace Microsoft.AspNetCore.Mvc
 
         [Theory]
         [MemberData(nameof(ValuesData))]
-        public async Task AcceptedAtActionResult_SetsStatusCodeAndValueAsync(object value)
+        public async Task AcceptedObjectResult_SetsStatusCodeAndValueAsync(object value)
         {
-            // Arrange           
-            var expectedUrl = "testAction";
+            // Arrange
+            var location = "/test/";
             var httpContext = GetHttpContext();
             var actionContext = GetActionContext(httpContext);
-            var urlHelper = GetMockUrlHelper(expectedUrl);
+            var result = new AcceptedResult(location, value);
 
             // Act
-            var result = new AcceptedAtActionResult(
-                actionName: expectedUrl,
-                controllerName: null,
-                routeValues: null,
-                value: value);
-
-            result.UrlHelper = urlHelper;
             await result.ExecuteResultAsync(actionContext);
 
             // Assert
@@ -83,49 +68,51 @@ namespace Microsoft.AspNetCore.Mvc
         }
 
         [Fact]
-        public async Task AcceptedAtActionResult_ReturnsStatusCode_SetsLocationHeader()
+        public void AcceptedObjectResult_SetsLocation()
         {
             // Arrange
-            var expectedUrl = "testAction";
-            var httpContext = GetHttpContext();
-            var actionContext = GetActionContext(httpContext);
-            var urlHelper = GetMockUrlHelper(expectedUrl);
+            var location = "http://test/location";
 
             // Act
-            var result = new AcceptedAtActionResult(
-                actionName: expectedUrl,
-                controllerName: null,
-                routeValues: null,
-                value: null);
+            var result = new AcceptedResult(location, "testInput");
 
-            result.UrlHelper = urlHelper;
+            // Assert
+            Assert.Same(location, result.Location);
+        }
+
+        [Fact]
+        public async Task AcceptedObjectResult_ReturnsStatusCode_SetsLocationHeader()
+        {
+            // Arrange
+            var location = "/test/";
+            var httpContext = GetHttpContext();
+            var actionContext = GetActionContext(httpContext);
+            var result = new AcceptedResult(location, "testInput");
+
+            // Act
             await result.ExecuteResultAsync(actionContext);
 
             // Assert
             Assert.Equal(StatusCodes.Status202Accepted, httpContext.Response.StatusCode);
-            Assert.Equal(expectedUrl, httpContext.Response.Headers["Location"]);
+            Assert.Equal(location, httpContext.Response.Headers["Location"]);
         }
 
         [Fact]
-        public async Task AcceptedAtActionResult_ThrowsOnNullUrl()
+        public async Task AcceptedObjectResult_OverwritesLocationHeader()
         {
             // Arrange
+            var location = "/test/";
             var httpContext = GetHttpContext();
             var actionContext = GetActionContext(httpContext);
-            var urlHelper = GetMockUrlHelper(returnValue: null);
+            httpContext.Response.Headers["Location"] = "/different/location/";
+            var result = new AcceptedResult(location, "testInput");
 
-            var result = new AcceptedAtActionResult(
-                actionName: null,
-                controllerName: null,
-                routeValues: null,
-                value: null);
+            // Act
+            await result.ExecuteResultAsync(actionContext);
 
-            result.UrlHelper = urlHelper;
-
-            // Act & Assert
-            await ExceptionAssert.ThrowsAsync<InvalidOperationException>(
-                async () => await result.ExecuteResultAsync(actionContext),
-            "No route matches the supplied values.");
+            // Assert
+            Assert.Equal(StatusCodes.Status202Accepted, httpContext.Response.StatusCode);
+            Assert.Equal(location, httpContext.Response.Headers["Location"]);
         }
 
         private static ActionContext GetActionContext(HttpContext httpContext)
@@ -137,6 +124,7 @@ namespace Microsoft.AspNetCore.Mvc
                                     routeData,
                                     new ActionDescriptor());
         }
+
         private static HttpContext GetHttpContext()
         {
             var httpContext = new DefaultHttpContext();
@@ -161,14 +149,6 @@ namespace Microsoft.AspNetCore.Mvc
                 NullLoggerFactory.Instance));
 
             return services.BuildServiceProvider();
-        }
-
-        private static IUrlHelper GetMockUrlHelper(string returnValue)
-        {
-            var urlHelper = new Mock<IUrlHelper>();
-            urlHelper.Setup(o => o.Action(It.IsAny<UrlActionContext>())).Returns(returnValue);
-
-            return urlHelper.Object;
         }
 
         private class ResourceStatusBody
