+++ test/Microsoft.AspNetCore.Mvc.FunctionalTests/WebApiCompatShimActionResultTest.cs
@@ -143,71 +143,6 @@ namespace Microsoft.AspNetCore.Mvc.FunctionalTests
         }
 
         [Fact]
-        public async Task ApiController_AcceptedRelative()
-        {
-            // Arrange & Act
-            var response = await Client.GetAsync("http://localhost/api/Blog/ActionResult/GetAcceptedRelative");
-            var content = await response.Content.ReadAsStringAsync();
-
-            // Assert
-            Assert.Equal(HttpStatusCode.Accepted, response.StatusCode);
-            Assert.Equal("{\"name\":\"Test User\"}", content);
-            Assert.Equal("5", response.Headers.Location.OriginalString);
-        }
-
-        [Fact]
-        public async Task ApiController_AcceptedAbsolute()
-        {
-            // Arrange & Act
-            var response = await Client.GetAsync("http://localhost/api/Blog/ActionResult/GetAcceptedAbsolute");
-            var content = await response.Content.ReadAsStringAsync();
-
-            // Assert
-            Assert.Equal(HttpStatusCode.Accepted, response.StatusCode);
-            Assert.Equal("{\"name\":\"Test User\"}", content);
-            Assert.Equal("/api/Blog/ActionResult/GetUser/5", response.Headers.Location.OriginalString);
-        }
-
-        [Fact]
-        public async Task ApiController_AcceptedQualified()
-        {
-            // Arrange & Act
-            var response = await Client.GetAsync("http://localhost/api/Blog/ActionResult/GetAcceptedQualified");
-            var content = await response.Content.ReadAsStringAsync();
-
-            // Assert
-            Assert.Equal(HttpStatusCode.Accepted, response.StatusCode);
-            Assert.Equal("{\"name\":\"Test User\"}", content);
-            Assert.Equal("http://localhost/api/Blog/ActionResult/5", response.Headers.Location.OriginalString);
-        }
-
-        [Fact]
-        public async Task ApiController_AcceptedUri()
-        {
-            // Arrange & Act
-            var response = await Client.GetAsync("http://localhost/api/Blog/ActionResult/GetAcceptedUri");
-            var content = await response.Content.ReadAsStringAsync();
-
-            // Assert
-            Assert.Equal(HttpStatusCode.Accepted, response.StatusCode);
-            Assert.Equal("{\"name\":\"Test User\"}", content);
-            Assert.Equal("/api/Blog/ActionResult/GetUser/5", response.Headers.Location.OriginalString);
-        }
-
-        [Fact]
-        public async Task ApiController_AcceptedAtRoute()
-        {
-            // Arrange & Act
-            var response = await Client.GetAsync("http://localhost/api/Blog/ActionResult/GetAcceptedAtRoute");
-            var content = await response.Content.ReadAsStringAsync();
-
-            // Assert
-            Assert.Equal(HttpStatusCode.Accepted, response.StatusCode);
-            Assert.Equal("{\"name\":\"Test User\"}", content);
-            Assert.Equal("http://localhost/api/Blog/ActionResult/GetUser/5", response.Headers.Location.OriginalString);
-        }
-
-        [Fact]
         public async Task ApiController_InternalServerError()
         {
             // Arrange & Act
